robz_meadows_backend/requirements.txt
asgiref==3.8.1
Django==4.2.23
djangorestframework==3.16.0
importlib_metadata==8.7.0
Markdown==3.8.2
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
zipp==3.23.0
